[Razor] Break down static web assets from the Razor SDK
This changes creates a new "internal" SDK for static web assets similar
to how the Web SDK is broken down internally in multiple SDKs.

This helps to clearly separate the set of targets/tools that deal with
compiling Razor files into assemblies from the targets/tools that deal
with static assets in the context of web applications and razor class
libraries.